+++ .travis.yml
@@ -6,6 +6,8 @@ node_js: "8"
 env:
   global:
     - _FORCE_LOGS=1
+  - COMMAND="npm test"
+  - COMMAND="npm run e2e-test" _FORCE_LOGS=1
 before_install:
   # Chrome Driver needs an updated version of Chrome
   - wget -P /tmp/ https://dl.google.com/chrome/mac/stable/GGRO/googlechrome.dmg
