+++ lib/chromedriver.js
@@ -9,14 +9,54 @@ import { SubProcess, exec } from 'teen_process';
 import B from 'bluebird';
 import { getChromedriverBinaryPath, getChromedriverDir } from './install';
 import { getChromeVersion } from './utils';
-import path from 'path';
 import semver from 'semver';
+import _ from 'lodash';
 
 
 const log = logger.getLogger('Chromedriver');
 
 const DEFAULT_HOST = '127.0.0.1';
 const DEFAULT_PORT = 9515;
+const CHROMEDRIVER_CHROME_MAPPING = {
+  '2.36': '63.0.3239',
+  '2.35': '62.0.3202',
+  '2.34': '61.0.3163',
+  '2.33': '60.0.3112',
+  '2.32': '59.0.3071',
+  '2.31': '58.0.3029',
+  '2.30': '58.0.3029',
+  '2.29': '57.0.2987',
+  '2.28': '55.0.2883',
+  '2.27': '54.0.2840',
+  '2.26': '53.0.2785',
+  '2.25': '53.0.2785',
+  '2.24': '52.0.2743',
+  '2.23': '51.0.2704',
+  '2.22': '49.0.2623',
+  '2.21': '46.0.2490',
+  '2.20': '43.0.2357',
+  '2.19': '43.0.2357',
+  '2.18': '43.0.2357',
+  '2.17': '42.0.2311',
+  '2.16': '42.0.2311',
+  '2.15': '40.0.2214',
+  '2.14': '39.0.2171',
+  '2.13': '38.0.2125',
+  '2.12': '36.0.1985',
+  '2.11': '36.0.1985',
+  '2.10': '33.0.1751',
+  '2.9': '31.0.1650',
+  '2.8': '30.0.1573',
+  '2.7': '30.0.1573',
+  // still trying to figure out the minimums for 2.6 and 2.5
+  '2.6': '',
+  '2.5': '',
+  '2.4': '29.0.1545',
+  '2.3': '28.0.1500',
+  '2.2': '27.0.1453',
+  '2.1': '27.0.1453',
+  '2.0': '27.0.1453',
+};
 
 class Chromedriver extends events.EventEmitter {
   constructor (args = {}) {
@@ -54,8 +94,7 @@ class Chromedriver extends events.EventEmitter {
       return await getChromedriverBinaryPath();
     }
 
-    const chromeVersion = (semver.coerce(await getChromeVersion(this.adb, this.bundleId)));
-    const mapping = JSON.parse(await fs.readFile(path.resolve(__dirname, '..', '..', 'versions.json')));
+    const chromeVersion = semver.coerce(await getChromeVersion(this.adb, this.bundleId));
 
     // go through the versions available
     const executables = await fs.glob(`${this.executableDir}/*`);
@@ -68,7 +107,7 @@ class Chromedriver extends events.EventEmitter {
           return {
             executable,
             version,
-            minCDVersion: mapping[version],
+            minCDVersion: CHROMEDRIVER_CHROME_MAPPING[version],
           };
         }
       } catch (ign) {}
@@ -81,9 +120,10 @@ class Chromedriver extends events.EventEmitter {
     const workingCds = cds.filter((cd) => semver.gte(chromeVersion, cd.minCDVersion))
       .sort((a, b) => a.version <= b.version ? 1 : -1);
 
-    if (!workingCds.length) {
+    if (_.isEmpty(workingCds)) {
       log.errorAndThrow(`No Chromedriver found that can automate Chrome '${chromeVersion}'.` +
-                        'See https://github.com/appium/appium/blob/master/docs/en/writing-running-appium/web/chromedriver.md for more details.');
+                        `See https://github.com/appium/appium/blob/master/docs/en/writing-running-appium/web/chromedriver.md ` +
+                        `for more details.`);
     }
 
     const binPath = workingCds[0].executable;
