+++ lib/utils.js
@@ -1,6 +1,6 @@
 async function getChromeVersion (adb, bundleId) {
-  const output = await adb.getPackageInfo(bundleId);
-  return output.versionName;
+  const {versionName} = await adb.getPackageInfo(bundleId);
+  return versionName;
 }
 
 export { getChromeVersion };
