+++ package.json
@@ -71,6 +71,7 @@
     "eslint-plugin-promise": "^3.3.1",
     "gulp": "^3.9.0",
     "mocha": "^5.0.5",
-    "pre-commit": "^1.1.3"
+    "pre-commit": "^1.1.3",
+    "sinon": "^4.4.8"
   }
 }
