+++ test/install-e2e-specs.js
