--- versions.json
@@ -1,39 +1,0 @@
-{
-  "2.36": "63.0.3239",
-  "2.35": "62.0.3202",
-  "2.34": "61.0.3163",
-  "2.33": "60.0.3112",
-  "2.32": "59.0.3071",
-  "2.31": "58.0.3029",
-  "2.30": "58.0.3029",
-  "2.29": "57.0.2987",
-  "2.28": "55.0.2883",
-  "2.27": "54.0.2840",
-  "2.26": "53.0.2785",
-  "2.25": "53.0.2785",
-  "2.24": "52.0.2743",
-  "2.23": "51.0.2704",
-  "2.22": "49.0.2623",
-  "2.21": "46.0.2490",
-  "2.20": "43.0.2357",
-  "2.19": "43.0.2357",
-  "2.18": "43.0.2357",
-  "2.17": "42.0.2311",
-  "2.16": "42.0.2311",
-  "2.15": "40.0.2214",
-  "2.14": "39.0.2171",
-  "2.13": "38.0.2125",
-  "2.12": "36.0.1985",
-  "2.11": "36.0.1985",
-  "2.10": "33.0.1751",
-  "2.9": "31.0.1650",
-  "2.8": "30.0.1573",
-  "2.7": "30.0.1573",
-  "2.6": "",
-  "2.5": "",
-  "2.4": "29.0.1545",
-  "2.3": "28.0.1500",
-  "2.2": "27.0.1453",
-  "2.1": "27.0.1453",
-  "2.0": "27.0.1453"
-}
